+++ backend/src/main/java/net/primepro/primepro/repository/AttendanceRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
-    List<Attendance> findByEmployeeId(Long id);
+public interface AttendanceRepository extends JpaRepository<Attendance, Integer> {
+    List<Attendance> findByEmployeeId(Integer id);
 }
