+++ backend/src/main/java/net/primepro/primepro/service/TaskService.java
@@ -8,11 +8,11 @@ import java.util.List;
 import java.util.Optional;
 
 public interface TaskService {
-    Task createTask(Long customerId, CarWashBookingDto bookingDto);
-    Optional<Task> getTaskById(Long taskId);
-    List<Task> getTasksByEmployee(Long employeeId);
-    Task updateTask(Long taskId, Task task);
-    void deleteTask(Long taskId);
-    Task assignEmployeeToTask(Long taskId, Long employeeId);
-    Task changeTaskStatus(Long taskId, BookingStatusEnum newStatus);
+    Task createTask(Integer customerId, CarWashBookingDto bookingDto);
+    Optional<Task> getTaskById(Integer taskId);
+    List<Task> getTasksByEmployee(Integer employeeId);
+    Task updateTask(Integer taskId, Task task);
+    void deleteTask(Integer taskId);
+    Task assignEmployeeToTask(Integer taskId, Integer employeeId);
+    Task changeTaskStatus(Integer taskId, BookingStatusEnum newStatus);
 }
